Scrape the hacks.youngrewiredstate.org site.
Unfortunately my first idea of searching github for YRS projects did not work, as there are many projects there with the text "YRS" that did not take part in the competition. Instead I am going to scrape the site for github project pages, store them in a database and write another script to download the data from that url.